requirements.txt
pandas==2.2.2
numpy==1.26.4
torch==2.3.0
ollama==0.2.0
streamlit==1.35.0
pinecone-client==4.1.0
sentence-transformers==3.0.0